Rename button to KnightHacksLinkButtons
