+++ src/common/KnightHacksLinkButtons.ts
